Investigating CDC SDK
