+++ config.go
@@ -3,6 +3,7 @@ package main
 const defaultLogLevel = "info"
 
 type cdcConfig struct {
+	mode           string
 	database       string
 	logAsJSON      bool
 	logLevel       string
