+++ go.mod
@@ -3,7 +3,7 @@ module github.com/radekg/yugabyte-db-cdctest
 go 1.16
 
 require (
-	github.com/hashicorp/go-hclog v0.16.2
-	github.com/radekg/yugabyte-db-go-client v0.0.2-beta.1
-	github.com/radekg/yugabyte-db-go-proto/v2 v2.11.2
+	github.com/hashicorp/go-hclog v1.2.0
+	github.com/radekg/yugabyte-db-go-client v0.0.2-beta.2
+	github.com/radekg/yugabyte-db-go-proto/v2 v2.13.0
 )
